General cleanup of ApplicationMessenger
Remove the circular dependency with CApplication
Move the DelayedThreadMessage to the cpp file as it's an implementation detail.
